orchestrators/floe-dagster/pyproject.toml
[project]
name = "floe_dagster"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster==1.12.13",
    "dagster-dg-cli>=1.12.13",
    "dagster-webserver>=1.12.13",
    "pyyaml>=6.0.0",
    "dg>=0.0.3",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "floe_dagster"
registry_modules = [
    "floe_dagster.components.*",
]

